+++ Z80/decoder.h
@@ -1,12 +1,12 @@
 #ifndef DECODER
 #define DECODER
 #include "Logger.h"
-#include "Z80/Memory.h"
-#include "Z80/alu.h"
+#include "Memory.h"
+#include "alu.h"
 
 class Decoder {
     Logger logger;
 public:
-    void decode(Memory& memory, Alu& alu,std::uint8_t& pc);
+    void decode(Memory& memory, Alu& alu,std::uint16_t& pc);
 };
 #endif
